refactor: 将 MCP 插件和 skills 功能重构为内置插件系统
- 将 MCP 插件和 skills 插件合并到 engine 的内置插件系统
- 删除独立的 mcp-plugin 和 skills 包
- 添加迁移文档说明变更
- 更新 package.json 和依赖配置
- 优化整体架构，减少包间依赖
