+++ packages/engine/package.json
@@ -17,6 +17,10 @@
     "./types": {
       "types": "./dist/types.d.ts",
       "import": "./dist/types.js"
+    },
+    "./built-in": {
+      "types": "./dist/built-in/index.d.ts",
+      "import": "./dist/built-in/index.js"
     }
   },
   "scripts": {
@@ -26,10 +30,12 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@ai-sdk/mcp": "^1.0.19",
     "@ai-sdk/openai": "^3.0.21",
     "ai": "^6.0.57",
     "dotenv": "^17.2.3",
     "glob": "^11.0.0",
+    "gray-matter": "^4.0.3",
     "yaml": "^2.8.2",
     "zod": "^4.3.6"
   },
