+++ packages/engine/src/Engine.ts
@@ -6,6 +6,7 @@ import type { UserConfigPluginLoadOptions } from './plugin/UserConfigPlugin.js';
 import { loop } from './core/loop.js';
 import { BuiltinToolsMap } from './tools/index.js';
 import { PluginManager } from './plugin/PluginManager.js';
+import { builtInPlugins } from './built-in/index.js';
 
 /**
  * 引擎配置选项
@@ -17,13 +18,16 @@ export interface EngineOptions {
   // 用户配置插件配置
   userConfigPlugins?: UserConfigPluginLoadOptions;
 
+  // 是否禁用内置插件（默认启用）
+  disableBuiltInPlugins?: boolean;
+
   // 全局配置
   config?: Record<string, any>;
 }
 
 /**
  * 重构后的引擎类
- * 支持双轨插件系统
+ * 自动包含内置插件，支持可选禁用
  */
 export class Engine {
   private pluginManager: PluginManager;
@@ -41,15 +45,19 @@ export class Engine {
 
   /**
    * 初始化引擎和插件系统
+   * 自动包含内置插件
    */
   async initialize(): Promise<void> {
     console.log('Initializing engine...', this.config);
 
+    // 准备插件列表：内置插件 + 用户配置插件
+    const allEnginePlugins = this.prepareEnginePlugins();
+
     // 插件管理器会自动处理加载顺序
     await this.pluginManager.initialize({
       enginePlugins: {
-        ...(this.options.enginePlugins || {})
-      }, // 将在loadPlugin时设置
+        ...allEnginePlugins
+      },
       userConfigPlugins: {
         ...(this.options.userConfigPlugins || {})
       }
@@ -58,6 +66,28 @@ export class Engine {
     // 合并插件工具到引擎工具库
     const pluginTools = this.pluginManager.getTools();
     this.tools = { ...this.tools, ...pluginTools };
+  }
+
+  /**
+   * 准备引擎插件列表（包含内置插件）
+   */
+  private prepareEnginePlugins(): EnginePluginLoadOptions {
+    const userPlugins = this.options.enginePlugins || {};
+    
+    // 如果用户禁用了内置插件，只返回用户插件
+    if (this.options.disableBuiltInPlugins) {
+      return userPlugins;
+    }
+
+    // 合并内置插件和用户插件
+    const builtInPluginList = [...builtInPlugins];
+    const userPluginList = userPlugins.plugins || [];
+
+    return {
+      plugins: [...builtInPluginList, ...userPluginList],
+      dirs: userPlugins.dirs || ['.coder/engine-plugins', '~/.coder/engine-plugins'],
+      scan: userPlugins.scan !== false // 默认启用扫描
+    };
   }
 
   /**
