+++ packages/engine/src/index.ts
@@ -5,6 +5,9 @@ export type { EnginePlugin, EnginePluginContext } from './plugin/EnginePlugin.js
 export type { UserConfigPlugin, UserConfigPluginLoadOptions } from './plugin/UserConfigPlugin.js';
 export { PluginManager } from './plugin/PluginManager.js';
 
+// 内置插件导出
+export { builtInPlugins, builtInMCPPlugin, builtInSkillsPlugin, BuiltInSkillRegistry } from './built-in/index.js';
+
 // 原有导出保持不变
 export * from './shared/types.js';
 export { loop } from './core/loop.js';
