--- packages/mcp-plugin/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@coder/mcp-plugin",
-  "version": "1.0.0",
-  "description": "Lightweight MCP plugin for Coder engine",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@ai-sdk/mcp": "1.0.19",
-    "@coder/engine": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^25.0.10",
-    "tsup": "^8.0.0",
-    "typescript": "^5.0.0"
-  },
-  "peerDependencies": {
-    "@coder/engine": "workspace:*"
-  }
-}
