--- packages/mcp-plugin/src/transport.ts
@@ -1,10 +1,0 @@
-export interface HTTPTransportConfig {
-  url: string;
-}
-
-export function createHTTPTransport(config: HTTPTransportConfig) {
-  return {
-    type: 'http' as const,
-    url: config.url
-  };
-}
