--- packages/skills/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@coder/skills",
-  "version": "1.0.0",
-  "description": "Skill system implementation for Coder",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
-    "test": "vitest",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@coder/engine": "workspace:*",
-    "glob": "^11.0.1",
-    "gray-matter": "^4.0.3",
-    "zod": "^4.3.6"
-  },
-  "devDependencies": {
-    "typescript": "^5.0.0",
-    "tsup": "^8.0.0",
-    "vitest": "^1.0.0",
-    "@types/node": "^25.0.10"
-  }
-}
