--- packages/skills/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm'],
-  dts: true,
-  clean: true,
-  splitting: false,
-  sourcemap: true,
-  target: 'es2022'
-});
